requirements.txt
Django
djangorestframework
djangorestframework-simplejwt
django-allauth
dj-rest-auth

requests
gunicorn